--- public/css/globals.css
@@ -1,10 +1,0 @@
-.copyToClipboard {
-    position:absolute;
-    top: 1em;
-    right: 1em;
-}
-
-.copyToClipboard:hover {
-    color: #106cfc;
-    cursor: pointer;
-}
