+++ src/views/index.ejs
@@ -11,33 +11,20 @@
     <meta name="theme-color" content="#ffffff">
     <link
       rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
     />
-    <link rel="stylesheet" href="/css/globals.css">
     <title>Elliot on Demand</title>
   </head>
   <body>
     <%- include('./partials/_nav'); %>
-    <h4 class="text-center m-3 p-3 bg-light text-dark">New episodes every day! Here's what we have on tap 🍺:</h4>
     <div class="container mt-2">
-      <div class="container px-4 mb-4">
-        <div class="row gx-5">
-          <div class="col">
-            <div class="overflow-auto p-3 bg-light" 
-              style="max-height: 80vh;">
-                <a class="twitter-timeline" data-theme="light" href="https://twitter.com/EITMonline?ref_src=twsrc%5Etfw">Tweets by EITMonline</a> <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
-              </div>
-          </div>
-          <div class="col">
-            <header>
-              <div id="now_playing" class="mt-3"></div>
-            </header>
-            <audio id="audioplayer" class="mt-2" controls autoplay>
-              <source src="" type="audio/wav">
-            </audio>
-          </div>
-        </div>
-      </div>
+      <header>
+        <h5 id="now_playing"></h5>
+      </header>
+      <audio id="audioplayer" controls autoplay>
+        <source src="" type="audio/wav">
+      </audio>
+      <p>New episodes every day! Here's what we have on tap 🍺:</p>
       <div class="list-group">
         <% episodes.forEach((item) => { %>
           <button 
@@ -57,7 +44,6 @@
         <a target="_blank" href="https://github.com/ethan-massey">Ethan</a>
       </footer>
     </div>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz" crossorigin="anonymous"></script>
     <script src="../js/episodeProgressFunctions.js" type="text/javascript"></script>
     <script>
       // inital setup
@@ -82,15 +68,8 @@
       // Also, add progress badge to each button
       var episodeButtons = document.getElementsByClassName("episodeName")
       for(var i = 0; i < episodeButtons.length; i++) {
-        episodeButtons[i].addEventListener("click", changeEpisode, false);
+        episodeButtons[i].addEventListener("click", changeCurrentEpisode, false);
         updateStatusBadge(episodeButtons[i].getAttribute('fileName'))
-      }
-
-      // If episode in URL, play that episode
-      const episodeQueued = <%- episodeQueued !== undefined ? JSON.stringify(episodeQueued) : 'undefined' %>;
-      if (episodeQueued !== undefined) {
-        episodeButton = document.getElementById(episodeQueued.fileName);
-        episodeButton.click();
       }
     </script>
   </body>
