+++ src/views/partials/_nav.ejs
@@ -1,5 +1,5 @@
 <nav class="navbar navbar-dark bg-primary navbar-expand">
-    <div class="container"><a class="navbar-brand" href="/">Elliot on Demand!</a>
+    <div class="container"><a class="navbar-brand" href="">Elliot on Demand!</a>
       <ul class="navbar-nav mr-auto">
         <!-- <li class="nav-item"><a class="nav-link" href="#"><span>Home</span></a></li>
         <li class="nav-item"><a class="nav-link" href="#"><span>About</span></a></li>
